requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn==0.24.0
gunicorn==21.2.0
starlette==0.27.0
pydantic==2.4.2
pydantic-settings==2.0.3
python-multipart==0.0.6
email-validator==2.1.0

# Database and Storage
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0
aiomysql==0.2.0
aiopg==1.4.0
pymongo==4.6.0
motor==3.3.1
redis==5.0.1
aioredis==2.0.1
influxdb-client==1.38.0
aioinflux==0.9.0

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-jwt==4.0.0
cryptography==41.0.5
certifi==2023.7.22
authlib==1.2.1

# Data Processing and Validation
numpy==1.26.2
pandas==2.1.3
pyyaml==6.0.1
jsonschema==4.19.2
marshmallow==3.20.1
python-dateutil==2.8.2

# Cache and Session Management
cachetools==5.3.2
aiocache==0.12.2
itsdangerous==2.1.2

# Metrics and Monitoring
prometheus-client==0.18.0
opentelemetry-api==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-instrumentation-fastapi==0.41b0
statsd==4.0.1
psutil==5.9.6

# Task Queue and Background Jobs
celery==5.3.5
redis==5.0.1
flower==2.0.1
aiohttp==3.9.1
asyncio==3.4.3

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.14
mkdocstrings==0.23.0
mkdocstrings-python==1.7.3

# Development Tools
black==23.11.0
flake8==6.1.0
mypy==1.7.1
isort==5.12.0
autoflake==2.2.1
pre-commit==3.5.0

# Utilities
python-dotenv==1.0.0
humanize==4.9.0
pytz==2023.3.post1
python-slugify==8.0.1
structlog==23.2.0
rich==13.7.0
typer==0.9.0
click==8.1.7
tqdm==4.66.1

# Container and Deployment
docker==6.1.3
kubernetes==27.2.0
boto3==1.29.3

# API Client and Networking
requests==2.31.0
aiohttp==3.9.1
httpx==0.25.2
websockets==12.0
urllib3==2.1.0

# Parsing and Formatting
beautifulsoup4==4.12.2
lxml==4.9.3
markdown==3.5.1
jinja2==3.1.2

# Performance and Optimization
ujson==5.8.0
orjson==3.9.10
msgpack==1.0.7
zstandard==0.22.0

# Profiling and Debugging
line-profiler==4.1.1
memory-profiler==0.61.0
debugpy==1.8.0

# Version Control Integration
gitpython==3.1.40

# Dependencies with Security Updates
setuptools>=65.5.1
wheel>=0.38.1